Move missing codec test to media_test
So it doesn't conflict with WASM.

Related to #449
