+++ peerconnection_test.go
@@ -2,13 +2,11 @@ package webrtc
 
 import (
 	"fmt"
-	"math/rand"
 	"reflect"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/pion/sdp/v2"
 	"github.com/pion/webrtc/v2/pkg/rtcerr"
 	"github.com/stretchr/testify/assert"
 )
@@ -315,48 +313,6 @@ func TestCreateOfferAnswer(t *testing.T) {
 	err = offerPeerConn.SetRemoteDescription(answer)
 	if err != nil {
 		t.Errorf("SetRemoteDescription (Originator): got error: %v", err)
-	}
-}
-
-func TestOfferRejectionMissingCodec(t *testing.T) {
-	api := NewAPI()
-	api.mediaEngine.RegisterDefaultCodecs()
-	pc, err := api.NewPeerConnection(Configuration{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	noCodecAPI := NewAPI()
-	noCodecPC, err := noCodecAPI.NewPeerConnection(Configuration{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	track, err := pc.NewTrack(DefaultPayloadTypeVP8, rand.Uint32(), "video", "pion2")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := pc.AddTrack(track); err != nil {
-		t.Fatal(err)
-	}
-
-	if err := signalPair(pc, noCodecPC); err != nil {
-		t.Fatal(err)
-	}
-
-	var sdes sdp.SessionDescription
-	if err := sdes.Unmarshal([]byte(pc.RemoteDescription().SDP)); err != nil {
-		t.Fatal(err)
-	}
-	var videoDesc sdp.MediaDescription
-	for _, m := range sdes.MediaDescriptions {
-		if m.MediaName.Media == "video" {
-			videoDesc = *m
-		}
-	}
-
-	if got, want := videoDesc.MediaName.Formats, []string{"0"}; !reflect.DeepEqual(got, want) {
-		t.Fatalf("rejecting unknown codec: sdp m=%s, want trailing 0", *videoDesc.MediaName.String())
 	}
 }
